fuzz: sometimes prefill some cmpctblock txns
This also changes the available set from being a set of transaction
indexes to a set of avaliable transaction wtxids:

The fuzz data provider sometimes generates blocks with duplicate
transactions (and txids). Previously, as no transactions were
prefilled, this would always result in a shortid-collision and
READ_STATUS_FAILED. However, since we now sometimes prefill one of
the duplicate transactions and provide a shortid for the other one,
no shortid-collision happens on duplicate transactions. This results
in READ_STATUS_OK and on READ_STATUS_OK we assert that for each
`IsTxAvailable()` transaction, we have an entry in the `avaliable` set.

Previously, we didn't add mark a transaction as `available` when the
txid already exists in `pool` and we might add only one of the duplicate
transactions to the extra pool and the other not. This resultes in
pdb.IsTxAvailable(i) being true while the transaction is not marked as
available. The following assertion would fail:

  assert(!pdb.IsTxAvailable(i) || available.count(i) > 0);

To fix this, we now mark wtxids as avaliable instead of of indexes.

An alternative would be to avoid fuzzing on blocks with duplicate
transactions. However, to better fuzz shortid collisions, we want
to have "duplicate" transactions in the block.